refact: melhorias, ajustes e inclusão de collections

+++ eazybytes/src/main/java/com/springsecurity/eazybytes/account/AccountController.java
@@ -14,13 +14,14 @@ public class AccountController {
 	private final AccountsRepository accountsRepository;
 
 	@GetMapping("/myAccount")
-	public Accounts getAccountDetails(@RequestParam Long id) {
+	public Accounts getAccountDetails(@RequestParam int id) {
 		Accounts accounts = accountsRepository.findByCustomerId(id);
-		if (accounts != null) {
+		if (accounts != null ) {
 			return accounts;
-		} else {
+		}else {
 			return null;
 		}
 	}
+
 
 }
