+++ eazybytes/src/main/java/com/springsecurity/eazybytes/account/AccountsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountsRepository extends JpaRepository<Accounts, Long> {
 
-	Accounts findByCustomerId(Long customerId);
+	Accounts findByCustomerId(int customerId);
 
 }
